Level-Select Menu Boosts
 - Adds an option the first time the game is run to either load from the save file or load from the SO's directly
 - On Andy's idea, scroll view container is now constrained to 5 columns (so you don't need to scroll horizontally)
 - Slightly cleaned up some functions in the levelManager script to work better.
 - Also created a new script "ButtonClickHandler".
 - The last two points are because:
      - after loading the scene the second time, buttons etc. became unlinked
      - could not find the script because the script is on an object in DoNotDestroy
      - new script works by finding the script in the DND area then calling the appropriate functions
      - Basically, so stupid that they were becoming unlinked, this is the best way I found to get around it.
